backend/requirements.txt
fastapi
uvicorn
minio
python-dotenv
sqlalchemy
python-multipart